Waiting for either android or java library plugin is applied before deciding the project type publishing strategy
