+++ buildSrc/src/main/java/co/elastic/apm/compile/tools/sourceheader/subplugins/JavaSourceHeaderPlugin.java
@@ -7,6 +7,6 @@ public class JavaSourceHeaderPlugin extends BaseSourceHeaderPlugin {
     @Override
     public void apply(Project project) {
         super.apply(project);
-        project.getTasks().getByName("classes", task -> task.dependsOn(getSpotlessApply(project)));
+        project.afterEvaluate(it -> it.getTasks().getByName("classes", task -> task.dependsOn(getSpotlessApply(it))));
     }
 }
